requirements.txt
aiogram>=3.6.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.30
python-dotenv>=1.0.1
jinja2>=3.1.4
python-multipart>=0.0.9
psycopg[binary]>=3.2.1
requests>=2.32.0
